Feat: started booking apps
Created an app for the booking system and started
a little with views, models and urls.

But first I´ve noticed that an error occured while
loading comments.js when there was no "deleteModal"
present. Which stopped users being able to log in.
This problem was solved with adding a code to
search for "deleteModal" before initializing the
function.