Feat: Start booking app
Started working with the booking system.
A lot of functions that need to work with
eachother. Came a little bit this session, need a
break from the coding now since there been smaller
problems everytime I configured some new function
into this booking system. Now I´ve got the right
data visual for the user when the booking form are
entered. Need to adjust so that the start_time and
end_time table fills in automaticly to the
database.

Note that there was a little problem when creating
this app. Needed to delete the whole database so
all the articles and about us etc was deleted. But
I´m focusing on fixing the booking system before
adding these articles again.