Feat: Create booking system
After several failures with finding the right
columns in the database, the booking system now
works and calculates the available times for the
upcoming 30days. Still have some more coding to do
with this system but it seems to be working with
booking available times and showing the user the
times they have booked. Now will I style the all
the html files and add all the articles again.
Because there was some trouble with the database
columns so I had to erase the database and do
the migration again.