Bug: Loading available times
When clicking on a court to make a booking there
are some problems to calculate the available times
for that court, first a booking was made and that
booked times where not available for any of the
dates ahead. Going to read a little about how to
fix that problem.