Feat: Booking system book times to database
After several tries to create a booking system,
now it´s working to generate and calculate times
that are available from today and 14days ahead.
Since there are 1h and 2h slots to book, the
function to fetch available times also check so
that the times don´t overlaps.

I´ve noticed in the early version of my booking
system that if the function itself generate
hours available it didn´t go that well. This time
I created a model with start_time and end_time
and there added which times a user can start and
end their booking with a maximum och 2 hours per
booking. This input of times made the system
easier to check available times for both courts to
be calculated.