Feat: Edit bookings
Added functions to edit and delete bookings that´s
been made before by a user. Tested with editing
and deleting a booking. So now the functions works
and there are some small touches to the booking
system left to do.

With this push I´ve got help to re-arrange the
view function from 2 functions to 1 function that
both Post and gets information from the database
and the javascripts calculation of available
times.

Next step with this booking system is to see if
there is a way to calculate away the available
timespots that has past what the time are when
a user are making a booking.