Feat: add filter to available times
Now the booking.js calculates on todays date and
time to show only available times that are after
the current time. So that no times that has past
the current time can be booked. Before when the
time was 9:25 am the time between '8:00 am - 9:00
am' was still available to book. With this filter
it disallow and not showing that time for the user
to book.

I´m really satisfied with the outcome of the
booking system, now there are some styling left to
do.