Style: Footer, rules and booking page styled
On the rules page I´ve changed the content text
about payment and accessment

On the booking page the divs with times and dates
now has a border and a "invisible" div between
them to seperate those divs a little.
Also added information to the user on how the
booking system works.

In the footer now are a copyright symbol, my
name and year added.