Fix: Data parsed to database fixed
Noticed that the function parsed time data in the
wrong format or nothing at all. So did some
debugging and found that the hidden divs that was
suppose to fetch the selected time and date didn´t
work along with the booking.js.
Tested after this and could make 3 confirmed
bookings on seperate dates.

Docs: wrote chapters about rules and booking pages