Docs: Lighthouse test
Ran the site through lighthouse performance test.
It noticed there was a lot build-time with having
the css/js-files loaded into the base.html.
Don´t think I need to change anything, since there
isn´t that much of a time to render it. With the
js-files loaded into base.html they will be in
"standby-mode" and loaded when they find an
element in their first if statement. The css-files
they refer to are mostly bootstraps own that are
needed.