Feat: Confirm comment done
Noticed that I didn´t have any confirmation when
a user submitting a comment, so basically copied
the modal for deleting an comment for the
structure. But instead of an eventlistener, the
view adding comment_posted to the URL that the
js-file are searching for to show the modal for
the user to confirm the post.