Fix: modal only shows when they should.
Noticed that the modals that should be shown when
the user successfully logged in or signed up
showed even if you filled in an invalid form when
signing up/logged in. Now there are functions
searching for URL parameters when signing up.
When logging in successfully it also shows up
correctly.