Docs: Manual test
Made some manual tests on the profile pages when
signing up, in and out with success. Made some
visual changes when testing. Changed "Log in"
to "Sign in" for an example.
Also tested all menu links on both small and big
screens.
And changed image in the readme file since it had
"Logged in as:" instead of "Signed in as:"