Fix: hide booking button
Ran into a problem that if a user tried to book or
update a time before they checked a available time
it shows a error that it´s wrong time format.
To fix that I added hidden class to the button and
an eventlistener that takes away the hidden class
when a radiobutton has been selected. To prevent
that error to happen.